Remove no needed v1 code and rename v2 to common file
