+++ plugins/wfp/management/commands/nigeria/Under5.py
@@ -6,7 +6,7 @@ from operator import itemgetter
 from django.core.paginator import Paginator
 
 from iaso.models import *
-from plugins.wfp.common_v2 import ETLV2
+from plugins.wfp.common import ETL
 from plugins.wfp.models import *
 
 
@@ -20,7 +20,7 @@ class NG_Under5:
 
     def run(self, updated_entity_ids, entity_type_code=None, task_name="etl_ng"):
         code = entity_type_code or self.ENTITY_TYPE_CODE
-        elt_v2 = ETLV2(code)
+        elt_v2 = ETL(code)
         account = elt_v2.get_account()
 
         page_size = self.PAGE_SIZE
@@ -32,7 +32,7 @@ class NG_Under5:
         )
 
         for page in pages:
-            submissions, page_info = ETLV2._retrieve_submissions(
+            submissions, page_info = ETL._retrieve_submissions(
                 code, updated_entity_ids, page_size=page_size, page_number=page
             )
             logger.info(f"Processing {len(page_info.object_list)} entities on page {page} for {account}")
