+++ plugins/wfp/management/commands/south_sudan/Under5.py
@@ -6,7 +6,7 @@ from operator import itemgetter
 from django.core.paginator import Paginator
 
 from iaso.models import Task
-from plugins.wfp.common_v2 import ETLV2
+from plugins.wfp.common import ETL
 from plugins.wfp.models import Beneficiary
 
 
@@ -47,7 +47,7 @@ class Under5:
             Name for the IASO Task log entry.
         """
         code = entity_type_code or self.ENTITY_TYPE_CODE
-        elt_v2 = ETLV2(code)
+        elt_v2 = ETL(code)
         account = elt_v2.get_account()
 
         page_size = self.PAGE_SIZE
@@ -58,7 +58,7 @@ class Under5:
             f"Processing {len(updated_entity_ids)} entities Child Under 5 across {paginator.num_pages} pages for {account}"
         )
         for page in pages:
-            submissions, page_info = ETLV2._retrieve_submissions(
+            submissions, page_info = ETL._retrieve_submissions(
                 code, updated_entity_ids, page_size=page_size, page_number=page
             )
             logger.info(f"Processing {len(page_info.object_list)} entities on page {page} for {account}")
